Cleaner and working version of reflection method added

+++ src/main/Position2D.java
@@ -1,18 +1,18 @@
 package main;
 
 public class Position2D {
-    private int x;
-    private int y;
-    Position2D(int x, int y){
+    private float x;
+    private float y;
+    Position2D(float x, float y){
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 }
